// types/color.types.ts
export type ThemeColors = {
    text: string;
    backgroundLight: string;
    backgroundDark: string;
    placeholderTextColor: string;
    inputBorderColor: string;
    inputBackgroundColor: string;
    buttonBackgroundColor: string;
    buttonTextColor: string;
    itemBackgroundColor: string;
    checkboxBorderColor: string;
    checkboxBackgroundColor: string;
    checkboxCompletedBackgroundColor: string;
    textCompletedColor: string;
    removeButtonTextColor: string;
};